Rename indexing utils
