+++ tutorials/Tutorial3_Basic_QA_Pipeline_without_Elasticsearch.py
@@ -11,7 +11,7 @@ from haystack import Finder
 from haystack.database.memory import InMemoryDocumentStore
 from haystack.database.sql import SQLDocumentStore
 from haystack.indexing.cleaning import clean_wiki_text
-from haystack.indexing.io import write_documents_to_db, fetch_archive_from_http
+from haystack.indexing.utils import convert_files_to_dicts, fetch_archive_from_http
 from haystack.reader.farm import FARMReader
 from haystack.reader.transformers import TransformersReader
 from haystack.retriever.tfidf import TfidfRetriever
@@ -37,11 +37,13 @@ doc_dir = "data/article_txt_got"
 s3_url = "https://s3.eu-central-1.amazonaws.com/deepset.ai-farm-qa/datasets/documents/wiki_gameofthrones_txt.zip"
 fetch_archive_from_http(url=s3_url, output_dir=doc_dir)
 
-# Now, let's write the docs to our DB.
+# convert files to dicts containing documents that can be indexed to our datastore
+dicts = convert_files_to_dicts(dir_path=doc_dir, clean_func=clean_wiki_text, split_paragraphs=True)
 # You can optionally supply a cleaning function that is applied to each doc (e.g. to remove footers)
 # It must take a str as input, and return a str.
-write_documents_to_db(document_store=document_store, document_dir=doc_dir, clean_func=clean_wiki_text, only_empty_db=True, split_paragraphs=True)
 
+# Now, let's write the docs to our DB.
+document_store.write_documents(dicts)
 
 
 # ## Initalize Retriever, Reader,  & Finder
